moved duplicate environment vars and getEnv.ts from tests folder to common code
