+++ tests/integration-tests/support/utils/aws/s3/s3Client.ts
@@ -1,4 +1,4 @@
 import { S3Client } from '@aws-sdk/client-s3'
-import { getEnv } from '../../getEnv'
+import { getEnv } from '../../../../../../common/utils/helpers/getEnv'
 
 export const s3Client = new S3Client({ region: getEnv('AWS_REGION') })
