--- tests/integration-tests/support/utils/getEnv.ts
@@ -1,10 +1,0 @@
-import { EnvironmentVar } from '../types/environmentVar'
-
-export const getEnv = (name: EnvironmentVar['name']) => {
-  const env = process.env[name]
-
-  if (env === undefined || env === null)
-    throw Error(`Missing environment variable: ${name}`)
-
-  return env
-}
